examples/prompt_registry/bridge/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "prompt-registry-bridge"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "langsmith>=0.1",
    "langchain-core>=0.3",
    "web3>=7.0",
    "eth-abi>=5.0",
    "python-dotenv>=1.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.setuptools.packages.find]
where = ["."]
